+++ CommonConcepts/Plugins/Rhetos.Dom.DefaultConcepts.Interfaces/InvalidDataMessage.cs
@@ -17,6 +17,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+using Rhetos.Dom.DefaultConcepts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -27,10 +28,42 @@ namespace Rhetos.Dom.DefaultConcepts
 {
     public class InvalidDataMessage
     {
-        public Guid ID;
         public string Message;
         public object[] MessageParameters = _emptyArray;
+        public Guid ID;
+        public IDictionary<string, object> Metadata;
+
+        public string Property { get { return (string)GetMetadata("Property"); } }
+        public string Validation { get { return (string)GetMetadata("Validation"); } }
+
+        public object GetMetadata(string key)
+        {
+            object value;
+            if (Metadata != null)
+                if (Metadata.TryGetValue(key, out value))
+                    return value;
+            return null;
+        }
 
         private static readonly object[] _emptyArray = new object[] { };
+
+        public static void ValidateOnSave(IEnumerable<IEntity> inserted, IEnumerable<IEntity> updated, IValidateRepository repository, string dataStructure)
+        {
+            if (inserted.Count() > 0 || updated.Count() > 0)
+            {
+                Guid[] newItemsIds = inserted.Concat(updated).Select(item => item.ID).ToArray();
+                var error = repository.Validate(newItemsIds, onSave: true).FirstOrDefault();
+                if (error != null)
+                {
+                    string systemMessage = "DataStructure:" + dataStructure + ",ID:" + error.ID.ToString()
+                        + (error.Metadata != null ? string.Concat(error.Metadata.Select(m => "," + m.Key + ":" + m.Value.ToString())) : "");
+
+                    throw new UserException(error.Message, error.MessageParameters, systemMessage, null)
+                    {
+                        Info = error.Metadata
+                    };
+                }
+            }
+        }
     }
 }