+++ CommonConcepts/Plugins/Rhetos.Dom.DefaultConcepts/Computations/InvalidDataCodeGenerator.cs
@@ -35,43 +35,50 @@ namespace Rhetos.Dom.DefaultConcepts
     [ExportMetadata(MefProvider.Implements, typeof(InvalidDataInfo))]
     public class InvalidDataCodeGenerator : IConceptCodeGenerator
     {
-        public static readonly CsTag<InvalidDataInfo> SystemMessageAppendTag = "SystemMessageAppend";
+        public static readonly CsTag<InvalidDataInfo> ErrorMetadataTag = "ErrorMetadata";
         public static readonly CsTag<InvalidDataInfo> OverrideUserMessagesTag = "OverrideUserMessages";
+
+        private readonly ConceptMetadata _conceptMetadata;
+
+        public InvalidDataCodeGenerator(ConceptMetadata conceptMetadata)
+        {
+            _conceptMetadata = conceptMetadata;
+        }
 
         public void GenerateCode(IConceptInfo conceptInfo, ICodeBuilder codeBuilder)
         {
             var info = (InvalidDataInfo)conceptInfo;
+            string dataStructure = info.Source.Module.Name + "." + info.Source.Name;
 
             // Using nonstandard naming of variables to avoid name clashes with injected code.
-            string getErrorMessageMethod =
-        @"public IEnumerable<InvalidDataMessage> " + info.GetErrorMessageMethodName() + @"(IEnumerable<Guid> invalidData_Ids)
-        {
-            const string invalidData_Description = " + CsUtility.QuotedString(info.ErrorMessage) + @";
-            " + OverrideUserMessagesTag.Evaluate(info) + @" return invalidData_Ids.Select(id => new InvalidDataMessage { ID = id, Message = invalidData_Description });
-        }
+            string errorMessageMethod =
+        $@"public IEnumerable<InvalidDataMessage> {info.GetErrorMessageMethodName()}(IEnumerable<Guid> invalidData_Ids)
+        {{
+            const string invalidData_Description = {CsUtility.QuotedString(info.ErrorMessage)};
+            IDictionary<string, object> metadata = new Dictionary<string, object>();
+            {ErrorMetadataTag.Evaluate(info)}
+            {OverrideUserMessagesTag.Evaluate(info)} return invalidData_Ids.Select(id => new InvalidDataMessage {{ ID = id, Message = invalidData_Description, Metadata = metadata }});
+        }}
 
         ";
-            codeBuilder.InsertCode(getErrorMessageMethod, RepositoryHelper.RepositoryMembers, info.Source);
+            codeBuilder.InsertCode(errorMessageMethod, RepositoryHelper.RepositoryMembers, info.Source);
             codeBuilder.AddReferencesFromDependency(typeof(InvalidDataMessage));
 
-            string dataStructure = info.Source.Module.Name + "." + info.Source.Name;
-            string systemMessage = @"""DataStructure:" + dataStructure + @",ID:"" + invalidItemId.ToString()" + SystemMessageAppendTag.Evaluate(info);
+            codeBuilder.InsertCode(
+                "metadata[\"Validation\"] = " + CsUtility.QuotedString(info.FilterType) + ";\r\n            ",
+                ErrorMetadataTag, info);
+
+            bool allowSave = _conceptMetadata.GetOrDefault(info, InvalidDataInfo.AllowSaveMetadata, false);
             string validationSnippet =
-                @"if (insertedNew.Count() > 0 || updatedNew.Count() > 0)
-                {
-                    Guid[] changedItemsId = inserted.Concat(updated).Select(item => item.ID).ToArray();
-                    Guid invalidItemId = this.Filter(this.Query(changedItemsId), new " + info.FilterType + @"())
-                        .Select(item => item.ID).FirstOrDefault();
-                    if (invalidItemId != default(Guid))
-                    {
-                        var userMessage = " + info.GetErrorMessageMethodName() + @"(new[] { invalidItemId }).Single();
-                        string systemMessage = " + systemMessage + @";
-                        throw new Rhetos.UserException(userMessage.Message, userMessage.MessageParameters, systemMessage, null);
-                    }
-                }
+                $@"if ({(allowSave ? "!" : "")}onSave)
+                {{
+                    var errorIds = this.Filter(this.Query(ids), new {info.FilterType}()).Select(item => item.ID).ToArray();
+                    if (errorIds.Count() > 0)
+                        foreach (var error in {info.GetErrorMessageMethodName()}(errorIds))
+                            yield return error;
+                }}
                 ";
-            codeBuilder.InsertCode(validationSnippet, WritableOrmDataStructureCodeGenerator.OnSaveTag2, info.Source);
-            codeBuilder.AddReferencesFromDependency(typeof(UserException));
+            codeBuilder.InsertCode(validationSnippet, WritableOrmDataStructureCodeGenerator.OnSaveValidateTag, info.Source);
         }
     }
 }