+++ CommonConcepts/Plugins/Rhetos.Dom.DefaultConcepts/Computations/InvalidDataMessageParametersItemCodeGenerator.cs
@@ -53,7 +53,8 @@ namespace Rhetos.Dom.DefaultConcepts
                 {
                     ID = parameters.ID,
                     Message = invalidData_Description,
-                    MessageParameters = " + parametersArray + @"
+                    MessageParameters = " + parametersArray + @",
+                    Metadata = metadata
                 });
             // ";
             codeBuilder.InsertCode(setMessages, InvalidDataCodeGenerator.OverrideUserMessagesTag, info.InvalidData);