+++ ChangeLog.md
@@ -1,5 +1,19 @@
 ﻿# Rhetos release notes
 
+## 1.6.0 (2017-02-27)
+
+### New features
+
+* New concept: **AllowSave**, for suppressing **InvalidData** validations on save.
+  The `Validate` method is available on the repository to list the validation errors that do not block saving.
+* New concept: **RequiredAllowSave**, for suppressing **Required** property validations on save.
+* New concept: **ErrorMetadata**, for setting the extensible metadata to **InvalidData** error message.
+
+### Internal improvements
+
+* Bugfix: The *SqlCommandTimeout* configuration option is not applied to the Entity Framework queries.
+* *AssemblyGenerator* uses UTF-8 with BOM when writing the source file (*ServerDom.cs*, for example).
+
 ## 1.5.0 (2017-02-13)
 
 ### New features